API: Finalize third batch of changes

+++ numpy/core/_add_newdocs.py
@@ -4581,7 +4581,7 @@ add_newdoc('numpy.core.multiarray', 'ndarray', ('view',
     .. note::
         Passing None for ``dtype`` is different from omitting the parameter,
         since the former invokes ``dtype(None)`` which is an alias for
-        ``dtype('float_')``.
+        ``dtype('float')``.
 
     Parameters
     ----------
